lucky_original_production/pyproject.toml
[project]
name = "botik-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioconsole>=0.8.2",
    "aiogram>=3.24.0",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "sqladmin>=0.22.0",
    "sqlalchemy>=2.0.46",
    "telethon>=1.42.0",
    "uvicorn>=0.40.0",
    "watchfiles>=1.1.1",
]

[dependency-groups]
dev = [
    "ty>=0.0.15",
]
